requirements.txt
--extra-index-url  https://download.pytorch.org/whl/rocm7.1
torch==2.10.0+rocm7.1
torchaudio==2.10.0+rocm7.1
torchmetrics==1.9.0
unbabel-comet==2.2.7
pandas
httpx
openpyxl
fastparquet
transformers
sacremoses
